Support mixed text and expressions in attributes
I originally thought this was unsupported and it made me sad.  I did it
again in a recent project update and it worked.  It also broke this
formatter.  So now we support these as arbitrarily nested.

As part of this work we also improve text node handling.  "External"
whitespace is collapsed to a single ASCII space.  We ignore unicode
whitespace on the theory that "they meant it."