Emit `raw` blocks from Parser
These tags have no semantic value in the AST, so they are
elided from that output.

Including them in the Parser output makes it possible to reconstruct the
source more correctly from said parse.  This is valuable when dealing
with the templates as source text, such as in the nascent Formatter
module.

In the AST builder, `render_code` is updated to `:skip` "raw" tags.
The most likely source of error here is handling a `[text: "raw"]` node.
Tests are added to demonstrate probable correctness.

This `:skip` arrangement makes it easy to add any future "source-only"
code handling, even though it is hard to see whence that would come.